Rename middleware to routes
