+++ services/sitemaps/src/routes/generate/format.js
