+++ services/sitemaps/src/routes/index.js
